Moved AsyncQueueTest to a separate file
