+++ AsyncCollections.Test/AsyncCollections.Test.csproj
@@ -78,6 +78,7 @@
     <Compile Include="AsyncBoundedPriorityQueueTest.cs" />
     <Compile Include="AsyncCollectionTakeFromAnyTest.cs" />
     <Compile Include="AsyncCollectionTest.cs" />
+    <Compile Include="AsyncQueueTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
